refactor(auth): extract status-gate predicates to shared side-effect-free subpath
Single source of truth at @classmoji/auth/predicates; server.ts keeps the
Response-throwing wrappers, apps/mcp keeps its ToolError wrappers, and the
parity test is no longer needed.
