+++ packages/auth/package.json
@@ -7,7 +7,8 @@
   },
   "exports": {
     "./server": "./src/server.ts",
-    "./client": "./src/client.ts"
+    "./client": "./src/client.ts",
+    "./predicates": "./src/predicates.ts"
   },
   "dependencies": {
     "@classmoji/database": "*",
